% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-subclustered_cells.R
\name{calculate_similarity}
\alias{calculate_similarity}
\title{Compute set similarity metrics}
\usage{
calculate_similarity(cluster_genes, ref_genes)
}
\arguments{
\item{cluster_genes}{Character vector. Genes found in the cluster.}

\item{ref_genes}{Character vector. Genes found in the reference signature.}
}
\value{
A list containing:
\itemize{
\item \code{jaccard}: Numeric value (0-1).
\item \code{overlap}: Numeric value (0-1).
}
}
\description{
Calculates the Jacquard index and overlap coefficient between two gene sets.
}
\details{
\strong{Jaccard Index:}
\deqn{J(A,B) = \frac{|A \cap B|}{|A \cup B|}}
Represents the intersection over the union.

\strong{Overlap Coefficient:}
\deqn{O(A,B) = \frac{|A \cap B|}{\min(|A|, |B|)}}
Represents the intersection relative to the size of the smaller set. This is often
more useful when comparing gene sets of vastly different sizes (e.g., a small
specific reference list vs a large cluster DE list).
}
