% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-markers.R
\name{.feature_space}
\alias{.feature_space}
\title{Define a complete feature-space}
\usage{
.feature_space(broad_marker_list, fine_marker_list)
}
\arguments{
\item{broad_marker_list}{A list of broad markers (simple or structured).}

\item{fine_marker_list}{A list of fine markers.}
}
\value{
A unique character vector of all features found in the inputs.
}
\description{
Extracts a unique vector of all features present in both the broad and fine marker lists.
Handles both structured (complex) and simple broad marker lists.
}
\keyword{internal}
