% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-seurat.R
\name{.is_norm_sObj}
\alias{.is_norm_sObj}
\title{Check for the presence of the Normalised Data Layer in a Seurat object.}
\usage{
.is_norm_sObj(sObj)
}
\arguments{
\item{sObj}{A Seurat object.}
}
\value{
NULL. Throws an error if the "data" layer is missing.
}
\description{
Verifies that the Seurat object contains a "data" layer (indicating normalisation has occurred).
}
\keyword{internal}
