% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-seurat.R
\name{.valid_sObj_input}
\alias{.valid_sObj_input}
\title{Validate Seurat object input}
\usage{
.valid_sObj_input(sObj)
}
\arguments{
\item{sObj}{The object to validate.}
}
\value{
NULL. Throws a an error if the object is not a Seurat object.
}
\description{
Checks if the provided object inherits from the "Seurat" class.
}
\keyword{internal}
