% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-markers.R
\name{feature_synonyms}
\alias{feature_synonyms}
\title{Expand gene symbols with synonyms from biomaRt}
\usage{
feature_synonyms(input_features, valid_genes = NULL, organism = "hsapiens")
}
\arguments{
\item{input_features}{A character vector of gene symbols OR a single regex pattern string.}

\item{valid_genes}{(Optional) A character vector of all gene names in your dataset.
If provided, used to filter results and allow regex matching. Default NULL.}

\item{organism}{Character. "hsapiens", "mmusculus", etc.}
}
\value{
A character vector of unique gene symbols found in \code{valid_genes}.
}
\description{
Connects to Ensembl to find synonyms for a list of genes. If \code{valid_genes} is provided,
it filters the results to match genes present in your dataset and supports regex patterns.
}
