% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-fisher_score.R
\name{get_best_match}
\alias{get_best_match}
\title{Get best assignments from Fisher scores}
\usage{
get_best_match(scores_df)
}
\arguments{
\item{scores_df}{Dataframe output from calculate_fisher_scores.}
}
\description{
Get best assignments from Fisher scores
}
