% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{nf}
\alias{nf}
\title{Create a new filename with a timestamp}
\usage{
nf(filename, filepath = NULL, timestamp_format = "\%Y-\%m-\%dT\%H-\%M-\%S")
}
\arguments{
\item{filename}{The filename.}

\item{filepath}{The path where the file should be saved.}

\item{timestamp_format}{The format for the timestamp}
}
\value{
a timestamped filename including a path (if one is supplied)
}
\description{
Create a new filename with a timestamp
}
