% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fisher_heatmap.R
\name{plot_fisher_heatmap}
\alias{plot_fisher_heatmap}
\title{Plot Fisher Score Heatmap}
\usage{
plot_fisher_heatmap(scores_df)
}
\arguments{
\item{scores_df}{Dataframe output from calculate_fisher_scores.}
}
\description{
Plot Fisher Score Heatmap
}
