% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{print_inline_vec}
\alias{print_inline_vec}
\title{Print an inline vector summary}
\usage{
print_inline_vec(counts, labels, text_color = "grey50", num_color = "grey90")
}
\arguments{
\item{counts}{Numeric vector. The count values.}

\item{labels}{Character vector. The names corresponding to the counts.}

\item{text_color}{Character. Color for the label text.}

\item{num_color}{Character. Color for the count numbers inside the brackets.}
}
\description{
Prints a list of labeled counts separated by vertical pipes.
Format: "label 1 (count 1) | label 2 (count 2) | ..."
}
