% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-broad_cluster.R
\name{run_broad_labelling}
\alias{run_broad_labelling}
\title{Perform initial broad cell type classification via Clustering}
\usage{
run_broad_labelling(
  object,
  broad_markers = list(Immune = "PTPRC", Endothelial = c("CDH5", "VWF"))
)
}
\arguments{
\item{object}{Seurat object.}

\item{broad_markers}{Named list of markers.}
}
\value{
Seurat object with 'Broad_Type' metadata.
}
\description{
Perform initial broad cell type classification via Clustering
}
