% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sc_expression_matrix.R
\name{save_seurat_components}
\alias{save_seurat_components}
\title{Extract and save Seurat counts and metadata to RDS}
\usage{
save_seurat_components(
  seurat_obj,
  metadata_cols = NULL,
  out_dir = getwd(),
  base_name = "data",
  assay = SeuratObject::DefaultAssay(seurat_obj),
  layer = "counts"
)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{metadata_cols}{A character vector of column names expected in metadata.
If NULL (default), all metadata columns are extracted.}

\item{out_dir}{Character. Directory where the files will be saved.}

\item{base_name}{Character. Prefix for the saved filenames.}

\item{assay}{Character. The specific assay to get data from or set data for; defaults to
\code{\link[SeuratObject]{DefaultAssay}}}

\item{layer}{The name of the layer to get (defaults to "count" for raw counts).}
}
\value{
Logical TRUE if successful.
}
\description{
Extract and save Seurat counts and metadata to RDS
}
