% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-fisher_score.R
\name{calculate_fisher_scores}
\alias{calculate_fisher_scores}
\title{Calculate Fisher's Exact Test Scores}
\usage{
calculate_fisher_scores(cluster_markers, reference_markers, background_genes)
}
\arguments{
\item{cluster_markers}{Dataframe with columns 'gene' and 'cluster' (from FindAllMarkers).}

\item{reference_markers}{Dataframe with 'gene' and 'cell_type' columns.}

\item{background_genes}{Character vector of all genes in the dataset.}
}
\description{
Calculate Fisher's Exact Test Scores
}
