% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-features.R
\name{feature_synonyms}
\alias{feature_synonyms}
\title{Expand gene symbols with synonyms from biomaRt}
\usage{
feature_synonyms(input_features, valid_genes, organism = "hsapiens")
}
\arguments{
\item{input_features}{A character vector of gene symbols OR a single regex pattern string.}

\item{valid_genes}{A character vector of all gene names in your dataset (to filter results).}

\item{organism}{Character. "hsapiens", "mmusculus", etc.}
}
\value{
A character vector of unique gene symbols found in \code{valid_genes}.
}
\description{
Connects to Ensembl to find synonyms for a list of genes or a regex pattern,
filtering for those actually present in your dataset.
}
