% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-cellvoter.R
\name{run_cellvoter}
\alias{run_cellvoter}
\title{Run the CellVoteR Labelling Method}
\usage{
run_cellvoter(
  object,
  method = 5,
  markers = NULL,
  panel_name = "idhwt_gbm_markers",
  triage_markers = NULL,
  verbose = TRUE,
  plot_heatmaps = FALSE
)
}
\arguments{
\item{object}{Input Seurat object.}

\item{method}{Integer (1-6). 5 & 6 are Global (Default). 1-4 are Triage.}

\item{markers}{User markers (or NULL to use internal).}

\item{panel_name}{If markers is NULL, which internal panel to use? Default "idhwt_gbm_markers".}

\item{triage_markers}{Named list of broad markers. If NULL, uses internal defaults.}

\item{verbose}{Logical.}

\item{plot_heatmaps}{Logical.}
}
\value{
List(object, heatmaps, scores)
}
\description{
Run the CellVoteR Labelling Method
}
